fix(#284): Post 수정 모달 안뜨는 버그 해결
